unit test cleanup
all tests now pass
