+++ expo/gbe/models/remains.py
@@ -1,4 +1,5 @@
 import os
+import pytz 
 from datetime import (
     datetime,
     timedelta
