+++ expo/gbe/models/volunteer.py
@@ -20,6 +20,7 @@ from expo.settings import (
     DATETIME_FORMAT,
     TIME_FORMAT,
 )
+from django.utils.formats import date_format
 
 
 class Volunteer(Biddable):
